VMS port: correct type err for context
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@9625 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/run-time/eif_vmsdirent.h
@@ -19,7 +19,7 @@ struct dirent {
     /* Handle returned by opendir(), used by the other routines.  You
      * are not supposed to care what's inside this structure. */
 typedef struct _dirdesc {
-    long			context;
+    unsigned int		context;
     int				vms_wantversions;
     char			*pattern;
     struct dirent		entry;
